+++ libraries/common/src/main/java/androidx/media3/common/util/Util.java
@@ -1863,7 +1863,9 @@ public final class Util {
         || encoding == C.ENCODING_PCM_16BIT
         || encoding == C.ENCODING_PCM_16BIT_BIG_ENDIAN
         || encoding == C.ENCODING_PCM_24BIT
+        || encoding == C.ENCODING_PCM_24BIT_BIG_ENDIAN
         || encoding == C.ENCODING_PCM_32BIT
+        || encoding == C.ENCODING_PCM_32BIT_BIG_ENDIAN
         || encoding == C.ENCODING_PCM_FLOAT;
   }
 
@@ -1876,7 +1878,9 @@ public final class Util {
   @UnstableApi
   public static boolean isEncodingHighResolutionPcm(@C.PcmEncoding int encoding) {
     return encoding == C.ENCODING_PCM_24BIT
+        || encoding == C.ENCODING_PCM_24BIT_BIG_ENDIAN
         || encoding == C.ENCODING_PCM_32BIT
+        || encoding == C.ENCODING_PCM_32BIT_BIG_ENDIAN
         || encoding == C.ENCODING_PCM_FLOAT;
   }
 
@@ -1992,8 +1996,10 @@ public final class Util {
       case C.ENCODING_PCM_16BIT_BIG_ENDIAN:
         return channelCount * 2;
       case C.ENCODING_PCM_24BIT:
+      case C.ENCODING_PCM_24BIT_BIG_ENDIAN:
         return channelCount * 3;
       case C.ENCODING_PCM_32BIT:
+      case C.ENCODING_PCM_32BIT_BIG_ENDIAN:
       case C.ENCODING_PCM_FLOAT:
         return channelCount * 4;
       case C.ENCODING_INVALID: