+++ libraries/exoplayer/src/main/java/androidx/media3/exoplayer/audio/ToFloatPcmAudioProcessor.java
@@ -29,7 +29,9 @@ import java.nio.ByteBuffer;
  *
  * <ul>
  *   <li>{@link C#ENCODING_PCM_24BIT}
+ *   <li>{@link C#ENCODING_PCM_24BIT_BIG_ENDIAN}
  *   <li>{@link C#ENCODING_PCM_32BIT}
+ *   <li>{@link C#ENCODING_PCM_32BIT_BIG_ENDIAN}
  *   <li>{@link C#ENCODING_PCM_FLOAT} ({@link #isActive()} will return {@code false})
  * </ul>
  */
@@ -70,6 +72,16 @@ import java.nio.ByteBuffer;
           writePcm32BitFloat(pcm32BitInteger, buffer);
         }
         break;
+      case C.ENCODING_PCM_24BIT_BIG_ENDIAN:
+        buffer = replaceOutputBuffer((size / 3) * 4);
+        for (int i = position; i < limit; i += 3) {
+          int pcm32BitInteger =
+              ((inputBuffer.get(i + 2) & 0xFF) << 8)
+                  | ((inputBuffer.get(i + 1) & 0xFF) << 16)
+                  | ((inputBuffer.get(i) & 0xFF) << 24);
+          writePcm32BitFloat(pcm32BitInteger, buffer);
+        }
+        break;
       case C.ENCODING_PCM_32BIT:
         buffer = replaceOutputBuffer(size);
         for (int i = position; i < limit; i += 4) {
@@ -78,6 +90,17 @@ import java.nio.ByteBuffer;
                   | ((inputBuffer.get(i + 1) & 0xFF) << 8)
                   | ((inputBuffer.get(i + 2) & 0xFF) << 16)
                   | ((inputBuffer.get(i + 3) & 0xFF) << 24);
+          writePcm32BitFloat(pcm32BitInteger, buffer);
+        }
+        break;
+      case C.ENCODING_PCM_32BIT_BIG_ENDIAN:
+        buffer = replaceOutputBuffer(size);
+        for (int i = position; i < limit; i += 4) {
+          int pcm32BitInteger =
+              (inputBuffer.get(i + 3) & 0xFF)
+                  | ((inputBuffer.get(i + 2) & 0xFF) << 8)
+                  | ((inputBuffer.get(i + 1) & 0xFF) << 16)
+                  | ((inputBuffer.get(i) & 0xFF) << 24);
           writePcm32BitFloat(pcm32BitInteger, buffer);
         }
         break;