+++ libraries/extractor/src/main/java/androidx/media3/extractor/WavUtil.java
@@ -77,7 +77,10 @@ public final class WavUtil {
         return TYPE_PCM;
       case C.ENCODING_PCM_FLOAT:
         return TYPE_FLOAT;
-      case C.ENCODING_PCM_16BIT_BIG_ENDIAN: // Not TYPE_PCM, because TYPE_PCM is little endian.
+        // TYPE_PCM is little endian so big endian formats don't match.
+      case C.ENCODING_PCM_16BIT_BIG_ENDIAN:
+      case C.ENCODING_PCM_24BIT_BIG_ENDIAN:
+      case C.ENCODING_PCM_32BIT_BIG_ENDIAN:
       case C.ENCODING_INVALID:
       case Format.NO_VALUE:
       default: